// session/getSessionRequestSchema.d.ts
export declare const getSessionRequestSchema: {
    type: string;
    properties: {
        sessionId: {
            type: string;
            description: string;
        };
    };
    required: string[];
};