// session/listSessionsResponseSchema.d.ts
export declare const listSessionsResponseSchema: {
    type: string;
    properties: {
        success: {
            type: string;
            description: string;
        };
        error: {
            type: string;
            description: string;
        };
        message: {
            type: string;
            description: string;
        };
        sessions: {
            type: string;
            items: {
                type: string;
            };
            description: string;
        };
    };
};
//# sourceMappingURL=listSessionsResponseSchema.d.ts.map